+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 - New SBCK.clim modules:
     - Add a the yearly_window function for BC
     - Add a function to apply a bias correction method
+    - Add a function to compute cross-auto-correlogram
+    - Add an optional function to compute cross-auto-correlogram with zxarray
 - New SBCK.misc modules:
     - Add linalg extension
     - Add sys extension